requirements.txt
image==1.5.33
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
Pillow==10.1.0
